requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.5
django-axes==5.40.1
django-cors-headers==4.1.0
django-ipware==5.0.0
django-password-validators==1.7.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.2
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.2.0
PyJWT==1.7.1
pytest==7.4.0
pytz==2023.3
requests==2.31.0
rest-api-payload==0.0.7
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==2.0.3
